requirements.txt
# Core Framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Resume Processing
PyMuPDF==1.23.8
python-docx==0.8.11
pdfplumber==0.9.0

# AI/ML Libraries
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
openai>=1.6.1,<2.0.0
chromadb==0.4.15
sentence-transformers==2.2.2

# Text Processing
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# Web Interface
streamlit==1.28.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
Pillow==10.1.0

# Development
pytest==7.4.3
pytest-flask==1.2.0
black==23.11.0
flake8==6.1.0
